refactor: 모달 tsx파일 이름 변경
